de_web_app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
pyspark==3.4.1
kafka-python==2.0.2
pydantic==1.10.11
python-dotenv==1.0.0
pytest==7.4.1
delta-spark==3.0.0
pyyaml==6.0
